new router: split into multiple files
