+++ src/user.php
@@ -1,0 +1,35 @@
+<?php
+
+class User {
+    public $address;
+    public $username;
+    public $domain;
+
+    public function isInvalid() {
+        global $config;
+        if (empty($this->username) || empty($this->domain)) {
+            return true;
+        } else if (!in_array($this->domain, $config['domains'])) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public static function parseDomain($address) {
+        $clean_address = _clean_address($address);
+        $user = new User();
+        $user->username = _clean_username($clean_address);
+        $user->domain = _clean_domain($clean_address);
+        $user->address = $user->username . '@' . $user->domain;
+        return $user;
+    }
+
+    public static function parseUsernameAndDomain($username, $domain) {
+        $user = new User();
+        $user->username = _clean_username($username);
+        $user->domain = _clean_domain($domain);
+        $user->address = $user->username . '@' . $user->domain;
+        return $user;
+    }
+}
